feat(album): add album viewer, selection bar, grid item, and meta panel components
- Implemented AlbumGridItem for displaying individual album thumbnails with selection and removal functionality.
- Created AlbumMetaPanel to show metadata for the currently viewed item, including EXIF data for images.
- Developed AlbumSelectionBar for managing multiple selections, including select all, clear selection, and delete options.
- Introduced AlbumViewer for viewing images and videos with zoom, pan, and video controls, along with metadata display.
